Move all of the dev-server files to a top-level dev-server directory
Make it easier to explore all the code and resources that make up the
development server by putting them all in one directory.
