+++ dev-server/templates/client-config.js.mustache
